Update 4  3/08/2023

+++ SoftwareMajorProject/notificationPage.cs
@@ -239,13 +239,13 @@ namespace SoftwareMajorProject
             string notificationLocation = txtNotificationLocation.Text;
             notificationDateTime = notificationDateSelected + " " +CmbNotificationHour.Text + ":" + CmbNotificationMinute.Text + ":00 " + CmbNotificationPeriod.Text;
 
-            MessageBox.Show(notificationDateTime);
+            //MessageBox.Show(notificationDateTime););*******************************************
 
 
 
             if (notificationDateSelected != "" && CmbNotificationHour.Text != "" && CmbNotificationMinute.Text != "" && CmbNotificationPeriod.Text != "")
             {
-                MessageBox.Show("Comparing dates");
+                //MessageBox.Show("Comparing dates"););*******************************************
 
                 DateTime dateTimeNow = DateTime.Now;
                 DateTime fullNotificationDateTime = Convert.ToDateTime(notificationDateTime);
@@ -261,7 +261,7 @@ namespace SoftwareMajorProject
                 }
                 else
                 {
-                    MessageBox.Show("Valid date!");
+                    MessageBox.Show("Notification added.");
                     
 
                     SQLiteConnection sqlConnection = new SQLiteConnection();
@@ -299,27 +299,36 @@ namespace SoftwareMajorProject
 
         private void btnDeleteNotification_Click(object sender, EventArgs e)
         {
-            SQLiteConnection sqlConnectionDeleteNotification = new SQLiteConnection();
-            sqlConnectionDeleteNotification.ConnectionString = "DataSource = softwareMajorProjectDatabase.db";
+            DialogResult checkDeleteNotification = MessageBox.Show("Are you sure you want to delete notification?", "", MessageBoxButtons.YesNo);
+            if (checkDeleteNotification == DialogResult.Yes)
+            {
+                
 
-            SQLiteCommand deleteNotificationCommand = new SQLiteCommand();
+                SQLiteConnection sqlConnectionDeleteNotification = new SQLiteConnection();
+                sqlConnectionDeleteNotification.ConnectionString = "DataSource = softwareMajorProjectDatabase.db";
 
-            deleteNotificationCommand.Connection = sqlConnectionDeleteNotification;
-            deleteNotificationCommand.CommandType = CommandType.Text;
-            deleteNotificationCommand.CommandText = "DELETE FROM Notifications WHERE NotificationIndex=" + selectedNotificationIndex;
+                SQLiteCommand deleteNotificationCommand = new SQLiteCommand();
 
-            deleteNotificationCommand.Parameters.AddWithValue("@NotificationIndex", selectedNotificationIndex);
+                deleteNotificationCommand.Connection = sqlConnectionDeleteNotification;
+                deleteNotificationCommand.CommandType = CommandType.Text;
+                deleteNotificationCommand.CommandText = "DELETE FROM Notifications WHERE NotificationIndex=" + selectedNotificationIndex;
 
-            sqlConnectionDeleteNotification.Open();
-            deleteNotificationCommand.ExecuteNonQuery();
-            sqlConnectionDeleteNotification.Close();
+                deleteNotificationCommand.Parameters.AddWithValue("@NotificationIndex", selectedNotificationIndex);
+
+                sqlConnectionDeleteNotification.Open();
+                deleteNotificationCommand.ExecuteNonQuery();
+                sqlConnectionDeleteNotification.Close();
 
 
-            //============================================================================================================================
+                LoadUserNotifications();
 
 
-            LoadUserNotifications();
-
+                MessageBox.Show("Notification deleted.");
+            }
+            else
+            {
+                MessageBox.Show("Notification not deleted");
+            }
         }
 
         private void DgvCurrentNotifications_SelectionChanged(object sender, EventArgs e)
@@ -373,11 +382,11 @@ namespace SoftwareMajorProject
                 int comparedDates = fullNotificationDateTime.CompareTo(dateTimeNow); //Compared the dateTime of the notification to the current dateTime
                 if (comparedDates < 0) //Occurs when the date is past current date
                 {
-                    MessageBox.Show(fullNotificationDateTime + " Has already passed");
+                    //MessageBox.Show(fullNotificationDateTime + " Has already passed");*************************************************
 
                     int i = 1;
 
-                    foreach (DataRow rowUserNotification in dataTableNotifications.Rows)
+                    //foreach (DataRow rowUserNotification in dataTableNotifications.Rows)
                     {
                         string checkIfNotificationOverdueCommand = "SELECT * FROM 'UserInfo'";
                         SQLiteDataAdapter notificationOverdueDataAdapter = new SQLiteDataAdapter(checkIfNotificationOverdueCommand, sqlConnectionNotificationsCheck);
@@ -391,7 +400,7 @@ namespace SoftwareMajorProject
                         sqlConnectionNotificationsCheck.Close();
 
 
-                        overdueUserName = rowUserNotification[1].ToString(); //Gets overdueUserName from the row in 'Notifications'
+                        overdueUserName = rowNotificationsCheck[1].ToString(); //Gets overdueUserName from the row in 'Notifications'
 
                         
 
@@ -400,7 +409,7 @@ namespace SoftwareMajorProject
                             if (rowNotificationOverdue[1].ToString() == overdueUserName && rowNotificationsCheck[1].ToString() == overdueUserName && i == 1)
                             {
                                 userEmail = rowNotificationOverdue[3].ToString(); //Gets userEmail from the row in 'UserInfo'
-                                overdueUserIndex = rowUserNotification[0].ToString();
+                                overdueUserIndex = rowNotificationsCheck[0].ToString();
 
                                 /*foreach (DataRow row in dataTableNotificationsCheck.Rows)
                                 {
@@ -412,11 +421,11 @@ namespace SoftwareMajorProject
                                     MessageBox.Show(userEmail + "<==== Here is an email.");
                                 }*/
 
-                                MessageBox.Show(userEmail + "<==== Here is an email." + fullNotificationDateTimeString);
+                                //MessageBox.Show(userEmail + "<==== Here is an email." + fullNotificationDateTimeString);*******************************************
 
 
-                                notificationTitleToSendToUser = "New Notification for " + overdueUserName + ": " + rowUserNotification[2].ToString();
-                                notificationBodyToSendToUser = "Title: " + rowUserNotification[2].ToString() + "\nDescription: " + rowUserNotification[3].ToString() + "\n Location: " + rowUserNotification[4].ToString() + "\n Time of Notification: " + rowUserNotification[5].ToString();
+                                notificationTitleToSendToUser = "New Notification for " + overdueUserName + ": " + rowNotificationsCheck[2].ToString();
+                                notificationBodyToSendToUser = "Title: " + rowNotificationsCheck[2].ToString() + "\nDescription: " + rowNotificationsCheck[3].ToString() + "\n Location: " + rowNotificationsCheck[4].ToString() + "\n Time of Notification: " + rowNotificationsCheck[5].ToString();
 
 
                                 MailMessage mailMessage = new MailMessage();
@@ -440,6 +449,8 @@ namespace SoftwareMajorProject
                                     MessageBox.Show("Failed to send. Error: " + ex.Message);
                                 }
 
+                                MessageBox.Show(overdueUserIndex);
+
 
                                 SQLiteConnection sqlConnectionDeleteNotification = new SQLiteConnection();
                                 sqlConnectionDeleteNotification.ConnectionString = "DataSource = softwareMajorProjectDatabase.db";
@@ -455,14 +466,7 @@ namespace SoftwareMajorProject
                                 sqlConnectionDeleteNotification.Close();
 
 
-
-
                                 LoadUserNotifications();
-
-
-                                //Prepare email to send to user about notification
-
-
 
                                 i = 0;
                             }
@@ -472,7 +476,7 @@ namespace SoftwareMajorProject
                 }
                 else if (comparedDates >= 0)
                 {
-                    MessageBox.Show(fullNotificationDateTime + " Date still to come");
+                    //MessageBox.Show(fullNotificationDateTime + " Date still to come");*******************************************
                 }
             }
